requirements.txt
discord.py
python-dotenv
gspread
google-auth
